Organizing s_package
